backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
pymongo==4.5.0
flask-jwt-extended==4.5.3
flask-mail==0.9.1
werkzeug==2.3.7
python-dotenv==1.0.0
requests==2.31.0
bcrypt==4.0.1
